review: address CodeRabbit + Copilot + Clara feedback on PR #368
Critical correctness:
- choose_npc_ability no longer skips abilities with required_ammo > 0.
  NPCs have infinite ammo by convention — the ammo check at the dispatch
  site is player-only. The previous selector logic would have
  permanently disabled NPC_DEFAULT_ABILITY (Pistol Shot 592, required
  ammo = 1 in seeded data) and held fire forever for every stock NPC.
  Regression guard pinned via selector_picks_ammo_bearing_ability_for_npc
  with the ability def seeded so the test fixture genuinely matches the
  bug shape (not just `ability_defs.get` returning None and falling
  through). Adversarial verification: re-introducing the buggy gate
  flips the guard to red (left: None, right: Some(592)).

Repo conventions:
- Strip issue refs (#329, #330) from Rust source comments and docs.
- "unspeced" → "unspecified" in spawn.rs comment and spawn-system.md.
- Test message "round-robin" → "alternate based on sort order" (the
  selector is deterministic, not RR).
- Correct the SQL comment in spawner/npcs.rs: actual query is a
  correlated subquery with COALESCE, not a LEFT JOIN with FILTER.

Test hardening:
- Chain 1032 replay test now asserts SetAggression target+level
  explicitly (was only asserting presence/order).
- selector_picks_ammo_bearing_ability_for_npc seeds ability_defs[592]
  with required_ammo=1 so it actually catches the regression.
- New selector_skips_multiple_cooling_abilities test exercises a
  3-ability bucket (Clara G8) — confirms the find walks past every
  cooling entry rather than stopping at the first.

Refactoring:
- Remove duplicated `aggression > 0` defensive check inside
  npc_ai_idle_auto_aggro; the outer npc_ai_tick filter is the single
  source of truth (repo convention: trust internal contracts).
- Trim narrative comments around choose_npc_ability — keep only the
  non-obvious "why" lines.
- Move inline `#[cfg(test)] mod tests` out of npc_ai.rs (now 368 lines,
  under the 500 soft cap) into the existing sibling test file at
  service/tests/npc_ai.rs.

Docs (Clara G3, G11):
- spawn-system.md notes cooldown state lives on
  AbilityManager::ability_cooldowns and is cleared by the leash tick.
- Adds a sentence that CellEntity::new initializes aggression=0, so
  respawn-eligible NPCs start passive again unless the mission chain
  re-fires set_aggression against the fresh spawn.

Clara feedback verified-and-correct (no fix needed):
- G1 (single source of truth): set_aggression already writes only to
  CellEntity::aggression; no property bag entry remains.
- G2 (selector fallback edge case): empty bucket → fallback,
  non-empty-but-all-gated → None. Distinct paths.
- G4 (1000 vs 1.0 dominance): numerical max_by guarantees this.
- G5 (JOIN multiplication): correlated subquery returns one row per
  spawn, no possible duplication.
- G7 (range gating placeholder): already documented in spawn-system.md.

Clara feedback skipped (with reason):
- G9 (party threat-dominance runtime test): chain-replay test already
  pins the 1000 magnitude; a full runtime party scenario would
  duplicate the threat_list infrastructure tests that already exist.
- G10 (byte-exact onAbilityLaunched(221)): the general
  onAbilityLaunched wire format is regression-guarded elsewhere; a
  per-ability byte test would be theatre.
