+++ crates/services/src/cell/spawner/npcs.rs
@@ -68,13 +68,14 @@ pub fn class_id_for_class(class: &str) -> u8 {
 pub async fn load_spawns_from_db(pool: &PgPool) -> Result<Vec<SpawnRecord>, sqlx::Error> {
     use sqlx::Row;
 
-    // LEFT JOIN + array_agg pulls the per-template ability bucket alongside
-    // the spawn row in one round-trip. `FILTER (WHERE ... IS NOT NULL)` keeps
-    // templates with no ability_set_id from materializing a `{NULL}` array.
-    // The COALESCE flips an empty (all-rows-filtered-out) aggregate to an
-    // empty Postgres array so Rust always sees `Vec<i32>` (possibly empty),
-    // never `None`. Templates without an ability set fall back to
-    // NPC_DEFAULT_ABILITY in `spawn_npc_from_record_into`.
+    // Correlated subquery + `COALESCE(..., ARRAY[]::int[])` pulls the
+    // per-template ability bucket alongside the spawn row in one round-trip.
+    // When the template's `ability_set_id` is NULL (no matching rows in
+    // `ability_set_abilities`), the inner `array_agg` returns NULL and
+    // COALESCE substitutes an empty Postgres array so Rust always sees
+    // `Vec<i32>` (possibly empty), never `None`. Templates without an
+    // ability set fall back to NPC_DEFAULT_ABILITY in
+    // `spawn_npc_from_record_into`.
     let rows = sqlx::query(
         "SELECT s.spawn_id, w.world AS world_name, s.x, s.y, s.z, s.heading, s.tag, \
                s.is_stationary, \
